WIP first compiling version ported to 1.13

+++ src/api/java/mezz/jei/api/gui/IAdvancedGuiHandler.java
@@ -4,9 +4,10 @@ import javax.annotation.Nullable;
 import java.awt.Rectangle;
 import java.util.List;
 
+import net.minecraft.client.gui.inventory.GuiContainer;
+
 import mezz.jei.api.IModRegistry;
 import mezz.jei.api.ingredients.IModIngredientRegistration;
-import net.minecraft.client.gui.inventory.GuiContainer;
 
 /**
  * Allows plugins to change how JEI is displayed next to their mod's guis.
@@ -31,9 +32,9 @@ public interface IAdvancedGuiHandler<T extends GuiContainer> {
 
 	/**
 	 * Return anything under the mouse that JEI could not normally detect, used for JEI recipe lookups.
-	 * <p>
+	 *
 	 * This is useful for guis that don't have normal slots (which is how JEI normally detects items under the mouse).
-	 * <p>
+	 *
 	 * This can also be used to let JEI look up liquids in tanks directly, by returning a FluidStack.
 	 * Works with any ingredient type that has been registered with {@link IModIngredientRegistration}.
 	 *
@@ -42,7 +43,7 @@ public interface IAdvancedGuiHandler<T extends GuiContainer> {
 	 * @since JEI 3.13.2
 	 */
 	@Nullable
-	default Object getIngredientUnderMouse(T guiContainer, int mouseX, int mouseY) {
+	default Object getIngredientUnderMouse(T guiContainer, double mouseX, double mouseY) {
 		return null;
 	}
 }
