+++ src/main/java/mezz/jei/gui/overlay/bookmarks/ILeftAreaContent.java
@@ -1,11 +1,12 @@
 package mezz.jei.gui.overlay.bookmarks;
 
-import mezz.jei.input.IShowsRecipeFocuses;
+import java.awt.Rectangle;
+import java.util.Set;
+
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.inventory.GuiContainer;
 
-import java.awt.Rectangle;
-import java.util.Set;
+import mezz.jei.input.IShowsRecipeFocuses;
 
 public interface ILeftAreaContent extends IShowsRecipeFocuses {
 
@@ -17,8 +18,8 @@ public interface ILeftAreaContent extends IShowsRecipeFocuses {
 
 	void updateBounds(Rectangle area, Set<Rectangle> guiExclusionAreas);
 
-	boolean handleMouseScrolled(int mouseX, int mouseY, int dWheel);
+	boolean handleMouseScrolled(double mouseX, double mouseY, double dWheel);
 
-	boolean handleMouseClicked(int mouseX, int mouseY, int mouseButton);
+	boolean handleMouseClicked(double mouseX, double mouseY, int mouseButton);
 
 }
