+++ src/main/java/mezz/jei/input/IShowsRecipeFocuses.java
@@ -5,7 +5,7 @@ import javax.annotation.Nullable;
 public interface IShowsRecipeFocuses {
 
 	@Nullable
-	IClickedIngredient<?> getIngredientUnderMouse(int mouseX, int mouseY);
+	IClickedIngredient<?> getIngredientUnderMouse(double mouseX, double mouseY);
 
 	boolean canSetFocusWithMouse();
